fix(debate): rename createDebateSession→createDebateRunner in tests; rename session-*.test.ts→runner-*.test.ts
Issue 1: Replace all _planDeps.createDebateSession references with
_planDeps.createDebateRunner in 6 plan-decompose test files, fixing
7 failing tests caused by the dep rename in src/cli/plan.ts.

Issue 2: Rename 9 session-*.test.ts files in test/unit/debate/ to
runner-*.test.ts to match the deleted session-*.ts source files.
For runner-stateful.test.ts and runner-plan.test.ts, merge coverage
from session-stateful and session-plan respectively, preserving all
unique tests without duplication.
