+++ test/unit/cli/plan-decompose-ac-repair.test.ts
@@ -83,7 +83,7 @@ const origWriteFile = _planDeps.writeFile;
 const origScanCodebase = _planDeps.scanCodebase;
 const origCreateManager = _planDeps.createManager;
 const origExistsSync = _planDeps.existsSync;
-const origCreateDebateSession = _planDeps.createDebateSession;
+const origCreateDebateRunner = _planDeps.createDebateRunner;
 const origDiscoverWorkspacePackages = _planDeps.discoverWorkspacePackages;
 const origReadPackageJson = _planDeps.readPackageJson;
 const origReadPackageJsonAt = _planDeps.readPackageJsonAt;
@@ -108,7 +108,7 @@ describe("planDecomposeCommand — AC overflow repair loop (issue #227)", () => 
     _planDeps.readPackageJsonAt = mock(async () => null);
     _planDeps.spawnSync = mock(() => ({ stdout: Buffer.from(""), exitCode: 1 }));
     _planDeps.mkdirp = mock(async () => {});
-    _planDeps.createDebateSession = mock(() => ({ run: mock(async () => ({ outcome: "failed" })) }) as never);
+    _planDeps.createDebateRunner = mock(() => ({ run: mock(async () => ({ outcome: "failed" })) }) as never);
   }
 
   beforeEach(async () => {
@@ -123,7 +123,7 @@ describe("planDecomposeCommand — AC overflow repair loop (issue #227)", () => 
     _planDeps.scanCodebase = origScanCodebase;
     _planDeps.createManager = origCreateManager;
     _planDeps.existsSync = origExistsSync;
-    _planDeps.createDebateSession = origCreateDebateSession;
+    _planDeps.createDebateRunner = origCreateDebateRunner;
     _planDeps.discoverWorkspacePackages = origDiscoverWorkspacePackages;
     _planDeps.readPackageJson = origReadPackageJson;
     _planDeps.readPackageJsonAt = origReadPackageJsonAt;
