+++ test/unit/cli/plan-decompose-guards.test.ts
@@ -103,7 +103,7 @@ const origWriteFile = _planDeps.writeFile;
 const origScanCodebase = _planDeps.scanCodebase;
 const origCreateManager = _planDeps.createManager;
 const origExistsSync = _planDeps.existsSync;
-const origCreateDebateSession = _planDeps.createDebateSession;
+const origCreateDebateRunner = _planDeps.createDebateRunner;
 const origDiscoverWorkspacePackages = _planDeps.discoverWorkspacePackages;
 const origReadPackageJson = _planDeps.readPackageJson;
 const origReadPackageJsonAt = _planDeps.readPackageJsonAt;
@@ -158,7 +158,7 @@ describe("planDecomposeCommand — guards (AC-1 to AC-8)", () => {
     _planDeps.scanCodebase = origScanCodebase;
     _planDeps.createManager = origCreateManager;
     _planDeps.existsSync = origExistsSync;
-    _planDeps.createDebateSession = origCreateDebateSession;
+    _planDeps.createDebateRunner = origCreateDebateRunner;
     _planDeps.discoverWorkspacePackages = origDiscoverWorkspacePackages;
     _planDeps.readPackageJson = origReadPackageJson;
     _planDeps.readPackageJsonAt = origReadPackageJsonAt;
