+++ test/unit/debate/runner-plan.test.ts
@@ -1,10 +1,13 @@
 import { afterEach, beforeEach, describe, expect, mock, test } from "bun:test";
+import type { NaxConfig } from "../../../src/config";
+import { DEFAULT_CONFIG } from "../../../src/config";
 import { DebateRunner } from "../../../src/debate/runner";
 import { _debateSessionDeps } from "../../../src/debate/session-helpers";
-import type { DebateStageConfig } from "../../../src/debate/types";
 import type { CallContext } from "../../../src/operations/types";
-import { DEFAULT_CONFIG } from "../../../src/config";
+import type { DebateStageConfig } from "../../../src/debate/types";
 import { makeMockAgentManager, makeSessionManager } from "../../helpers";
+
+// ─── Fixtures ─────────────────────────────────────────────────────────────────
 
 function makeCallCtx(overrides: Partial<CallContext> = {}): CallContext {
   const agentManager = makeMockAgentManager();
@@ -25,6 +28,28 @@ function makeCallCtx(overrides: Partial<CallContext> = {}): CallContext {
   };
 }
 
+function makeCallCtxWithIds(
+  storyId: string,
+  agentManager: ReturnType<typeof makeMockAgentManager>,
+  sessionManager: ReturnType<typeof makeSessionManager>,
+  config: NaxConfig = DEFAULT_CONFIG,
+): CallContext {
+  return {
+    runtime: {
+      agentManager,
+      sessionManager,
+      configLoader: { current: () => config, select: (_sel: unknown) => config } as any,
+      packages: { resolve: () => ({ config, select: (_sel: unknown) => config }) } as any,
+      signal: undefined,
+    } as any,
+    packageView: { config, select: (_sel: unknown) => config } as any,
+    packageDir: "/tmp/work",
+    agentName: "claude",
+    storyId,
+    featureName: "test",
+  };
+}
+
 function makePlanStageConfig(overrides: Partial<DebateStageConfig> = {}): DebateStageConfig {
   return {
     enabled: true,
@@ -39,6 +64,10 @@ function makePlanStageConfig(overrides: Partial<DebateStageConfig> = {}): Debate
     ...overrides,
   };
 }
+
+const TEST_CONFIG = {
+  autoMode: { defaultAgent: "opencode" },
+} as unknown as NaxConfig;
 
 let origGetSafeLogger: typeof _debateSessionDeps.getSafeLogger;
 let origReadFile: typeof _debateSessionDeps.readFile;
@@ -60,6 +89,8 @@ afterEach(() => {
   _debateSessionDeps.readFile = origReadFile;
   mock.restore();
 });
+
+// ─── Core plan mode tests ─────────────────────────────────────────────────────
 
 describe("DebateRunner.runPlan() — plan mode uses sessionManager.runInSession", () => {
   test("plan mode calls sessionManager.runInSession (not agentManager.planAs)", async () => {
@@ -110,9 +141,7 @@ describe("DebateRunner.runPlan() — plan mode uses sessionManager.runInSession"
       outputDir: "/tmp/out",
     });
 
-    // sessionManager.runInSession should be called per debater
     expect(runInSessionCalls.length).toBeGreaterThan(0);
-    // planAs should NOT be called in the migrated implementation
     expect(planAsCalls.length).toBe(0);
   });
 
@@ -197,7 +226,6 @@ describe("DebateRunner.runPlan() — plan mode uses sessionManager.runInSession"
       outputDir: "/tmp/out",
     });
 
-    // Should return a valid DebateResult shape
     expect(result).toHaveProperty("outcome");
     expect(result).toHaveProperty("stage");
     expect(result).toHaveProperty("storyId");
@@ -227,8 +255,6 @@ describe("DebateRunner.runPlan() — plan mode uses sessionManager.runInSession"
       sessionManager: undefined,
     });
 
-    // Each debater throws internally (no sessionManager) but allSettledBounded catches it
-    // and all debaters fail → buildFailedResult → outcome: "failed"
     const result = await runner.runPlan("task context", "output format", {
       workdir: "/tmp/work",
       feature: "feat-plan",
@@ -236,5 +262,389 @@ describe("DebateRunner.runPlan() — plan mode uses sessionManager.runInSession"
     });
 
     expect(result.outcome).toBe("failed");
+  });
+});
+
+// ─── Extended plan mode tests (from session-plan) ────────────────────────────
+
+describe("DebateRunner.runPlan()", () => {
+  test("passes unique indexed role to each plan debater via sessionManager.runInSession", async () => {
+    const sessionRoles: string[] = [];
+
+    const sm = makeSessionManager({
+      runInSession: mock(async (name: string, _prompt: string, opts: any) => {
+        sessionRoles.push(opts?.role ?? "");
+        return {
+          output: "ok",
+          tokenUsage: { inputTokens: 0, outputTokens: 0 },
+          internalRoundTrips: 0,
+        };
+      }) as any,
+      nameFor: mock((req: any) => `nax-${req?.role ?? "unknown"}`),
+    });
+
+    _debateSessionDeps.readFile = mock(async (path: string) => `output:${path}`);
+
+    const config = { ...TEST_CONFIG, debate: { enabled: true, agents: 3, maxConcurrentDebaters: 3 } } as unknown as NaxConfig;
+    const agentManager = makeMockAgentManager();
+
+    const runner = new DebateRunner({
+      ctx: makeCallCtxWithIds("config-ssot", agentManager, sm, config),
+      stage: "plan",
+      stageConfig: makePlanStageConfig({
+        debaters: [{ agent: "opencode" }, { agent: "opencode" }, { agent: "opencode" }],
+      }),
+      config,
+      workdir: "/tmp/workdir",
+      sessionManager: sm,
+    });
+
+    await runner.runPlan("task context", "output format", {
+      workdir: "/tmp/workdir",
+      feature: "config-ssot",
+      outputDir: "/tmp/out",
+    });
+
+    expect(sessionRoles).toHaveLength(3);
+    expect(sessionRoles[0]).toBe("plan-0");
+    expect(sessionRoles[1]).toBe("plan-1");
+    expect(sessionRoles[2]).toBe("plan-2");
+  });
+
+  test("passes storyId through to each sessionManager.runInSession call", async () => {
+    const capturedStoryIds: string[] = [];
+
+    const sm = makeSessionManager({
+      runInSession: mock(async (_name: string, _prompt: string, opts: any) => {
+        capturedStoryIds.push(opts?.storyId ?? "");
+        return {
+          output: "ok",
+          tokenUsage: { inputTokens: 0, outputTokens: 0 },
+          internalRoundTrips: 0,
+        };
+      }) as any,
+      nameFor: mock((req: any) => `nax-${req?.role ?? "unknown"}`),
+    });
+
+    _debateSessionDeps.readFile = mock(async () => "{}");
+
+    const config = { ...TEST_CONFIG, debate: { enabled: true, agents: 2, maxConcurrentDebaters: 2 } } as unknown as NaxConfig;
+    const agentManager = makeMockAgentManager();
+
+    const runner = new DebateRunner({
+      ctx: makeCallCtxWithIds("config-ssot", agentManager, sm, config),
+      stage: "plan",
+      stageConfig: makePlanStageConfig(),
+      config,
+      workdir: "/tmp/workdir",
+      sessionManager: sm,
+    });
+
+    await runner.runPlan("task context", "output format", {
+      workdir: "/tmp/workdir",
+      feature: "config-ssot",
+      outputDir: "/tmp/out",
+    });
+
+    expect(capturedStoryIds).toEqual(["config-ssot", "config-ssot"]);
+  });
+
+  test("runs hybrid rebuttal loop when mode=hybrid and sessionMode=stateful", async () => {
+    const rebuttalCalls: Array<{ prompt: string; handleId: string }> = [];
+    const closedHandleIds: string[] = [];
+
+    const mockSM = makeSessionManager({
+      openSession: mock(async (name: string) => ({ id: name, agentName: "opencode" })),
+      closeSession: mock(async (handle: any) => { closedHandleIds.push(handle.id); }),
+      nameFor: mock((req: any) => req?.role ?? ""),
+      runInSession: mock(async (_name: string, _prompt: string) => ({
+        output: "ok",
+        tokenUsage: { inputTokens: 0, outputTokens: 0 },
+        internalRoundTrips: 0,
+      })) as any,
+    });
+
+    const agentManager = makeMockAgentManager({
+      runAsSessionFn: async (_agentName, handle, prompt) => {
+        if (prompt.includes("You are debater") && prompt.includes("## Your Task")) {
+          rebuttalCalls.push({ prompt, handleId: handle.id });
+        }
+        return {
+          output: `run-output-${handle.id}`,
+          tokenUsage: { inputTokens: 0, outputTokens: 0 },
+          internalRoundTrips: 0,
+        };
+      },
+    });
+
+    _debateSessionDeps.readFile = mock(async (path: string) => `{"proposal":"from ${path}"}`);
+
+    const config = { ...TEST_CONFIG, debate: { enabled: true, agents: 2, maxConcurrentDebaters: 2 } } as unknown as NaxConfig;
+
+    const runner = new DebateRunner({
+      ctx: makeCallCtxWithIds("plan-hybrid-test", agentManager, mockSM, config),
+      stage: "plan",
+      stageConfig: makePlanStageConfig({
+        mode: "hybrid",
+        sessionMode: "stateful",
+        rounds: 1,
+        debaters: [{ agent: "opencode" }, { agent: "opencode" }],
+      }),
+      config,
+      workdir: "/tmp/workdir",
+      sessionManager: mockSM,
+    });
+
+    const result = await runner.runPlan("task context", "output format", {
+      workdir: "/tmp/workdir",
+      feature: "plan-hybrid-test",
+      outputDir: "/tmp/out",
+    });
+
+    expect(rebuttalCalls).toHaveLength(2);
+    expect(rebuttalCalls[0]?.handleId).toBe("plan-hybrid-0");
+    expect(rebuttalCalls[1]?.handleId).toBe("plan-hybrid-1");
+    expect(closedHandleIds).toHaveLength(2);
+    expect(result.rebuttals).toBeDefined();
+    expect(result.rebuttals).toHaveLength(2);
+    expect(result.rounds).toBe(1);
+  });
+
+  test("skips rebuttal loop when mode is panel (default)", async () => {
+    const runAsSessionCalls: Array<{ prompt: string }> = [];
+
+    const sm = makeSessionManager({
+      runInSession: mock(async () => ({
+        output: "ok",
+        tokenUsage: { inputTokens: 0, outputTokens: 0 },
+        internalRoundTrips: 0,
+      })) as any,
+      nameFor: mock((req: any) => `nax-${req?.role ?? "unknown"}`),
+    });
+
+    const agentManager = makeMockAgentManager({
+      runAsSessionFn: async (_agentName, _handle, prompt) => {
+        runAsSessionCalls.push({ prompt });
+        return {
+          output: "run-output",
+          tokenUsage: { inputTokens: 0, outputTokens: 0 },
+          internalRoundTrips: 0,
+        };
+      },
+    });
+
+    _debateSessionDeps.readFile = mock(async () => "{}");
+
+    const config = { ...TEST_CONFIG, debate: { enabled: true, agents: 2, maxConcurrentDebaters: 2 } } as unknown as NaxConfig;
+
+    const runner = new DebateRunner({
+      ctx: makeCallCtxWithIds("plan-panel-test", agentManager, sm, config),
+      stage: "plan",
+      stageConfig: makePlanStageConfig(),
+      config,
+      workdir: "/tmp/workdir",
+      sessionManager: sm,
+    });
+
+    const result = await runner.runPlan("task context", "output format", {
+      workdir: "/tmp/workdir",
+      feature: "plan-panel-test",
+      outputDir: "/tmp/out",
+    });
+
+    expect(runAsSessionCalls).toHaveLength(0);
+    expect(result.rebuttals).toBeUndefined();
+    expect(result.rounds).toBe(1);
+  });
+
+  test("warns and skips rebuttal when mode=hybrid but sessionMode=one-shot", async () => {
+    const warnings: string[] = [];
+    _debateSessionDeps.getSafeLogger = mock(() => ({
+      warn: (_stage: string, msg: string) => warnings.push(msg),
+      info: () => {},
+      debug: () => {},
+      error: () => {},
+    })) as unknown as typeof _debateSessionDeps.getSafeLogger;
+
+    const sm = makeSessionManager({
+      runInSession: mock(async () => ({
+        output: "ok",
+        tokenUsage: { inputTokens: 0, outputTokens: 0 },
+        internalRoundTrips: 0,
+      })) as any,
+      nameFor: mock((req: any) => `nax-${req?.role ?? "unknown"}`),
+    });
+
+    _debateSessionDeps.readFile = mock(async () => "{}");
+
+    const config = { ...TEST_CONFIG, debate: { enabled: true, agents: 2, maxConcurrentDebaters: 2 } } as unknown as NaxConfig;
+    const agentManager = makeMockAgentManager();
+
+    const runner = new DebateRunner({
+      ctx: makeCallCtxWithIds("plan-hybrid-oneshot", agentManager, sm, config),
+      stage: "plan",
+      stageConfig: makePlanStageConfig({
+        mode: "hybrid",
+        sessionMode: "one-shot",
+        rounds: 2,
+      }),
+      config,
+      workdir: "/tmp/workdir",
+      sessionManager: sm,
+    });
+
+    const result = await runner.runPlan("task context", "output format", {
+      workdir: "/tmp/workdir",
+      feature: "plan-hybrid-oneshot",
+      outputDir: "/tmp/out",
+    });
+
+    expect(warnings.some((w) => w.includes("hybrid") && w.includes("stateful"))).toBe(true);
+    expect(result.rebuttals).toBeUndefined();
+    expect(result.rounds).toBe(1);
+  });
+
+  test("includes spec anchor in synthesis prompt when specContent is provided", async () => {
+    let capturedSynthesisPrompt = "";
+
+    const sm = makeSessionManager({
+      runInSession: mock(async () => ({
+        output: "ok",
+        tokenUsage: { inputTokens: 0, outputTokens: 0 },
+        internalRoundTrips: 0,
+      })) as any,
+      nameFor: mock((req: any) => `nax-${req?.role ?? "unknown"}`),
+    });
+
+    const agentManager = makeMockAgentManager({
+      completeFn: async (_agentName, prompt) => {
+        capturedSynthesisPrompt = prompt;
+        return { output: '{"userStories":[]}', costUsd: 0, source: "fallback" as const };
+      },
+    });
+
+    _debateSessionDeps.readFile = mock(async () => '{"userStories":[]}');
+
+    const specContent = `# My Feature\n## Stories\n### US-001\n**AC:**\n- AC one\n- AC two`;
+    const config = { ...TEST_CONFIG, debate: { enabled: true, agents: 2, maxConcurrentDebaters: 2 } } as unknown as NaxConfig;
+
+    const runner = new DebateRunner({
+      ctx: makeCallCtxWithIds("spec-anchor-test", agentManager, sm, config),
+      stage: "plan",
+      stageConfig: makePlanStageConfig({
+        resolver: { type: "synthesis", agent: "opencode" },
+      }),
+      config,
+      workdir: "/tmp/workdir",
+      sessionManager: sm,
+    });
+
+    await runner.runPlan("task context", "output format", {
+      workdir: "/tmp/workdir",
+      feature: "spec-anchor-test",
+      outputDir: "/tmp/out",
+      specContent,
+    });
+
+    expect(capturedSynthesisPrompt).toContain("## Original Spec");
+    expect(capturedSynthesisPrompt).toContain("AC one");
+    expect(capturedSynthesisPrompt).toContain("AC two");
+    expect(capturedSynthesisPrompt).toContain("acceptanceCriteria");
+    expect(capturedSynthesisPrompt).toContain("suggestedCriteria");
+  });
+
+  test("synthesis prompt omits spec anchor when specContent is not provided", async () => {
+    let capturedSynthesisPrompt = "";
+
+    const sm = makeSessionManager({
+      runInSession: mock(async () => ({
+        output: "ok",
+        tokenUsage: { inputTokens: 0, outputTokens: 0 },
+        internalRoundTrips: 0,
+      })) as any,
+      nameFor: mock((req: any) => `nax-${req?.role ?? "unknown"}`),
+    });
+
+    const agentManager = makeMockAgentManager({
+      completeFn: async (_agentName, prompt) => {
+        capturedSynthesisPrompt = prompt;
+        return { output: '{"userStories":[]}', costUsd: 0, source: "fallback" as const };
+      },
+    });
+
+    _debateSessionDeps.readFile = mock(async () => '{"userStories":[]}');
+
+    const config = { ...TEST_CONFIG, debate: { enabled: true, agents: 2, maxConcurrentDebaters: 2 } } as unknown as NaxConfig;
+
+    const runner = new DebateRunner({
+      ctx: makeCallCtxWithIds("no-spec-anchor", agentManager, sm, config),
+      stage: "plan",
+      stageConfig: makePlanStageConfig({
+        resolver: { type: "synthesis", agent: "opencode" },
+      }),
+      config,
+      workdir: "/tmp/workdir",
+      sessionManager: sm,
+    });
+
+    await runner.runPlan("task context", "output format", {
+      workdir: "/tmp/workdir",
+      feature: "no-spec-anchor",
+      outputDir: "/tmp/out",
+    });
+
+    expect(capturedSynthesisPrompt).not.toContain("## Original Spec");
+    expect(capturedSynthesisPrompt).not.toContain("suggestedCriteria");
+  });
+
+  test("runs plan debaters in parallel (when limit >= agents)", async () => {
+    const startedOrder: number[] = [];
+    const resolvers: Array<() => void> = [];
+
+    const sm = makeSessionManager({
+      runInSession: mock(async (_name: string, _prompt: string, opts: any) => {
+        const index = Number((opts?.role ?? "").replace("plan-", ""));
+        startedOrder.push(index);
+        await new Promise<void>((resolve) => {
+          resolvers[index] = resolve;
+        });
+        return {
+          output: "ok",
+          tokenUsage: { inputTokens: 0, outputTokens: 0 },
+          internalRoundTrips: 0,
+        };
+      }) as any,
+      nameFor: mock((req: any) => `nax-${req?.role ?? "unknown"}`),
+    });
+
+    _debateSessionDeps.readFile = mock(async () => "{}");
+
+    const config = { ...TEST_CONFIG, debate: { enabled: true, agents: 2, maxConcurrentDebaters: 2 } } as unknown as NaxConfig;
+    const agentManager = makeMockAgentManager();
+
+    const runner = new DebateRunner({
+      ctx: makeCallCtxWithIds("config-ssot", agentManager, sm, config),
+      stage: "plan",
+      stageConfig: makePlanStageConfig(),
+      config,
+      workdir: "/tmp/workdir",
+      sessionManager: sm,
+    });
+
+    const runPromise = runner.runPlan("task context", "output format", {
+      workdir: "/tmp/workdir",
+      feature: "config-ssot",
+      outputDir: "/tmp/out",
+    });
+
+    for (let attempt = 0; attempt < 20; attempt++) {
+      if (startedOrder.length >= 2) break;
+      await Promise.resolve();
+    }
+    expect(startedOrder).toEqual([0, 1]);
+
+    resolvers[0]?.();
+    resolvers[1]?.();
+    await runPromise;
   });
 });
